# submission/api/openapi.yaml
openapi: 3.1.0
info:
  title: LabelHub Core API
  version: "1.0.0"
  summary: HTTP edge of the LabelHub data-annotation platform.
  description: >
    Portable OpenAPI description of LabelHub's **HTTP route handlers**
    (`src/app/api/**`). These thin external edges authenticate, then delegate to
    the library layer. They cover the **machine-to-machine** surfaces:
    workspace-API-key ingest/read endpoints, admin (Supabase session) export
    endpoints, public operational endpoints, and the LLM gateway proxy.


    LabelHub deliberately keeps **internal mutations off HTTP** — the browser UI
    creates tasks, submits/reviews annotations, and runs payouts through
    **Next.js Server Actions** (`'use server'` functions), NOT these routes. That
    RPC surface is not expressible in OpenAPI; it is documented in the companion
    `API.md` ("Server-Action RPC surface"). This file documents only what is
    reachable over HTTP.


    ## Authentication models

    Three distinct models are in play (see `securitySchemes`):

    1. **Workspace API key** — machine-to-machine. A `lh_ws_*` (or rate-limited
       public `lh_demo_*`) bearer token. Only the SHA-256 hash is stored; the
       workspace is *inferred from the key*, so these endpoints take no
       `workspaceId`. Accepted on `Authorization: Bearer <key>`, bare
       `Authorization: <key>`, `x-api-key: <key>`, or `x-labelhub-api-key: <key>`.
    2. **Supabase user session** — cookie session from `@supabase/ssr`. Used by
...

...
servers:
...
paths:
...
  /api/export/dataset:
    get:
      tags: [Export]
      summary: Export a frozen dataset version
      description: >
        Streams the frozen manifest of a dataset version. `format` controls the
        content shape (`raw` | `teaching`), `encoding` the file format
        (`json` | `jsonl` | `csv` | `excel`). Large exports (estimated > 5 MB)
        enqueue an async job and return 202 with a `statusUrl` to poll.
        Admin of the *version's* workspace (resolved before the guard).
      operationId: exportDatasetVersion
      security:
        - supabaseSession: []
      parameters:
        - name: versionId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: format
          in: query
          schema: { type: string, enum: [raw, teaching], default: raw }
        - name: encoding
          in: query
          schema: { type: string, enum: [json, jsonl, csv, excel], default: jsonl }
        - name: mapping
          in: query
          description: JSON array of `{ source, target, transform? }` field-mapping entries.
          schema: { type: string }
      responses:
        "200":
          description: Synchronous export stream (small datasets).
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
            application/json: {}
            application/jsonl: {}
            text/csv: {}
        "202":
          description: Async export enqueued; poll `statusUrl`.
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId: { type: string, format: uuid }
                  status: { type: string, const: pending }
                  statusUrl: { type: string }
                  estimatedBytes: { type: integer }
                  rowCount: { type: integer }
        "400":
          description: Missing `versionId` or unknown format/encoding (`VALIDATION_ERROR`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FlatError" }
        "401":
          description: Not signed in.
        "403":
          description: Not an admin of the version's workspace.
        "404":
          description: Dataset version not found (`NOT_FOUND`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FlatError" }
        "500":
          $ref: "#/components/responses/FlatInternalError"
  /api/export/jobs/{id}:
    get:
      tags: [Export]
      summary: Poll an async export job
      description: >
        Returns the job status; when `completed`, lazily mints a `downloadUrl`
        (signed Supabase URL, 60s TTL, or local-FS public URL). Admin of the
        job's workspace; 404 to non-admins (no existence leak).
      operationId: getExportJob
      security:
        - supabaseSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Job status.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  status: { type: string, enum: [pending, running, completed, failed] }
                  format: { type: string }
                  byteSize: { type: integer, nullable: true }
                  rowCount: { type: integer, nullable: true }
                  downloadUrl: { type: string, nullable: true }
                  error: { type: string, nullable: true }
                  createdAt: { type: string, format: date-time }
                  finishedAt: { type: string, format: date-time, nullable: true }
        "404":
          description: Job not found / not admin (`NOT_FOUND`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NestedError" }
        "500":
          $ref: "#/components/responses/NestedInternalError"
...
  /api/admin/compute-hints:
    post:
      tags: [Admin ops]
      summary: Synchronously compute trajectory hints
      description: >
        Token-gated. Runs the trajectory-hint reviewer inside the request
        lifetime (vs the async after()-window path) for retry-by-curl / demo
        pre-seeding. `GET` also works for browser pasting.
      operationId: adminComputeHints
      security:
        - adminToken: []
      parameters:
        - name: trajectoryId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Hint compute result (ok true/false).
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "400":
          description: Missing `trajectoryId` (`MISSING_PARAM`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NestedError" }
        "403":
          description: Bad token.
        "500":
          description: Compute failed (`INTERNAL`).
        "503":
          description: Admin endpoints disabled.
    get:
      tags: [Admin ops]
      summary: Synchronously compute trajectory hints (GET alias)
      operationId: adminComputeHintsGet
      security:
        - adminToken: []
      parameters:
        - name: trajectoryId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Hint compute result.
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
...
components:
  securitySchemes:
    workspaceApiKey:
      type: http
      scheme: bearer
      description: >
        Workspace API key (`lh_ws_*`, or rate-limited public `lh_demo_*`).
        Accepted on `Authorization: Bearer <key>`, bare `Authorization: <key>`,
        `x-api-key: <key>`, or `x-labelhub-api-key: <key>`. The workspace is
        inferred from the key.
    supabaseSession:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: >
        Supabase cookie session (`@supabase/ssr`). The route resolves the target
        workspace, then enforces a workspace-scoped guard
        (`requireWorkspaceAdmin` / `requireWorkspaceMember`). Cookie name varies
        by Supabase project ref.
    adminToken:
      type: http
      scheme: bearer
      description: >
        Out-of-band ops secret from the `ADMIN_DIAG_TOKEN` env. Send via
        `Authorization: Bearer <token>` or `x-admin-token`; `?token=` query is a
        soft-deprecated fallback.
  schemas:
    FlatError:
      type: object
      description: Error envelope used by ingest / export / eval-run routes.
      properties:
        error: { type: string }
        code: { type: string }
      required: [error, code]
    NestedError:
      type: object
      description: Error envelope used by the API-key read + admin routes.
      properties:
        error:
          type: object
          properties:
            message: { type: string }
            code: { type: string }
      required: [error]
...
  responses:
    FlatInternalError:
      description: "Internal error (generic; no stack/DB text). `code: INTERNAL`."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/FlatError" }
    NestedInternalError:
      description: "Internal error (generic; no stack/DB text). `code: INTERNAL`."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/NestedError" }
...